# Use handler directly instead of slog